broke the javascript file into two files based on what page the user is looking at. Also created a recipe class and test file fore it
