+++ index.html
@@ -266,7 +266,7 @@
                 OverlayScrollbars(document.querySelectorAll("body"), {});
             });
         </script>
-        <script src="./assets/java.js"></script>
+        <script src="./assets/landingHTML.js"></script>
 </body>
 
 </html>
