+++ recipes.html
@@ -62,8 +62,8 @@
   <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
     integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
     crossorigin="anonymous"></script>
-  <script async type="text/javascript" src="../js/bulma.js"></script>
-  <script src="./assets/java.js"></script>
+  <!-- <script async type="text/javascript" src="../js/bulma.js"></script> -->
+  <script src="./assets/recipeHTML.js"></script>
 </body>
 
 </html>
